requirements.txt
guardrails-ai==0.4.2
pydantic>=1.10.0,<2.0.0
typing-extensions>=4.0.0
requests>=2.25.0
numpy>=1.21.0
transformers>=4.20.0
torch>=1.12.0
scikit-learn>=1.0.0 